# server/docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: 若依 Go 版本接口文档
  title: RuoYi-Go API
  version: "1.0"
paths:
...
  /system/role:
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 新增角色
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 修改角色
      tags:
      - 角色管理
  /system/role/{id}:
    get:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 角色信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 查询角色详细
      tags:
      - 角色管理
  /system/role/{ids}:
    delete:
      parameters:
      - description: 角色ID，多个用逗号分隔
        in: path
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - 角色管理
  /system/role/authUser/allocatedList:
    get:
      parameters:
      - description: 角色ID
        in: query
        name: roleId
        required: true
        type: integer
      - description: 用户名称
        in: query
        name: userName
        type: string
      - description: 手机号码
        in: query
        name: phonenumber
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 查询角色已授权用户列表
      tags:
      - 角色管理
  /system/role/authUser/cancel:
    put:
      consumes:
      - application/json
      parameters:
      - description: 包含roleId和userId
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 取消用户授权角色
      tags:
      - 角色管理
  /system/role/authUser/cancelAll:
    put:
      consumes:
      - application/json
      parameters:
      - description: 包含roleId和userIds
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 批量取消用户授权角色
      tags:
      - 角色管理
  /system/role/authUser/selectAll:
    put:
      consumes:
      - application/json
      parameters:
      - description: 包含roleId和userIds
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 授权用户选择
      tags:
      - 角色管理
  /system/role/authUser/unallocatedList:
    get:
      parameters:
      - description: 角色ID
        in: query
        name: roleId
        required: true
        type: integer
      - description: 用户名称
        in: query
        name: userName
        type: string
      - description: 手机号码
        in: query
        name: phonenumber
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 查询角色未授权用户列表
      tags:
      - 角色管理
  /system/role/changeStatus:
    put:
      consumes:
      - application/json
      parameters:
      - description: 包含id和status
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 角色状态修改
      tags:
      - 角色管理
  /system/role/dataScope:
    put:
      consumes:
      - application/json
      parameters:
      - description: 数据权限信息
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 角色数据权限
      tags:
      - 角色管理
  /system/role/deptTree/{id}:
    get:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 部门树
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 根据角色ID查询部门树结构
      tags:
      - 角色管理
  /system/role/export:
    post:
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Excel文件
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 导出角色数据
      tags:
      - 角色管理
  /system/role/list:
    get:
      parameters:
      - description: 角色名称
        in: query
        name: roleName
        type: string
      - description: 权限字符
        in: query
        name: roleKey
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 角色列表
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 查询角色列表
      tags:
      - 角色管理
...
securityDefinitions:
  BearerAuth:
    description: Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"